+++ src/gui.js
@@ -91,7 +91,7 @@ modules.gui = '2026-September-22';
 
 // Declarations
 
-var SnapVersion = '12.1.4-dev';
+var SnapVersion = '12.2.0-dev';
 
 var IDE_Morph;
 var ProjectDialogMorph;