+++ HISTORY.md
@@ -2,8 +2,15 @@
 
 ## in development:
 
+## 12.2.0:
+* **New Features:**
+    * ability to hide sprites in the corral via the corral-bar's context menu, use for pedagogical puzzles
+* **Translation Updates:**
+    * German
+
 ### 2026-09-22
 * objects, gui, store: added ability to hide sprites in the corral via the corral-bar's context menu, use for pedagogical puzzles
+* incremented dev version to v12.2
 
 ### 2026-09-21
 * German translation update